requirements.txt
attrdict==2.0.1
base58==1.0.3
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
Django==2.2.3
django-cors-headers==2.2.0
django-rest-swagger==2.1.2
django-stubs==1.0.2
django-summernote==0.8.8.6
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.0.0
djoser==1.1.5
idna==2.8
inflection==0.3.1
ipfshttpclient==0.4.12
itypes==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
lru-dict==1.1.6
MarkupSafe==1.1.1
multiaddr==0.0.8
mypy==0.720
mypy-extensions==0.4.1
netaddr==0.7.19
openapi-codec==1.3.2
parsimonious==0.8.1
protobuf==3.9.0
pycryptodome==3.8.2
PyJWT==1.7.1
pytz==2019.1
requests==2.22.0
rlp==1.1.0
semantic-version==2.6.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
toolz==0.10.0
typed-ast==1.4.0
typing-extensions==3.7.4
uritemplate==3.0.0
urllib3==1.25.3
varint==1.0.2
web3==5.0.0
websockets==7.0
zope.interface==4.6.0
